Добавил загрузку картинок на сервер Делаю выгрузку изображений

+++ database/migrations/2019_03_30_183442_create_files_table.php
@@ -18,11 +18,12 @@ public function up()
     $table->string('name')->unique();
     $table->string('type');
     $table->string('extension');
-    $table->integer('user_id')->unsigned();
+    $table->bigInteger('user_id')->unsigned();
     $table->timestamps();
 
     $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
   });
+    
 }
 
     /**
